requirements.txt
# Core dependencies
playwright==1.40.0
google-cloud-bigquery==3.13.0  # 将来のBigQuery移行のために保持
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4

# AI/ML APIs
openai==1.6.1
google-generativeai==0.3.2
pinecone-client==3.0.2
tavily-python==0.3.0

# Google APIs
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data processing
pandas==2.3.0
pandas-gbq==0.19.2  # 将来のBigQuery移行のために保持
pydantic==2.5.2
openpyxl==3.1.2
numpy==1.26.2

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.25.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Database
supabase==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
pyjwt==2.10.1
email-validator==2.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
tenacity==8.2.3
python-dateutil==2.8.2
pyyaml==6.0.1
click==8.1.7

# Web scraping utilities
beautifulsoup4==4.12.2
lxml==4.9.3

# Task queue (for background jobs)
celery==5.3.4
redis==5.0.1

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.0
isort==5.13.2
mypy==1.7.1
pre-commit==3.6.0

# Type stubs
types-requests==2.31.0.10
pandas-stubs==2.2.3.250527